backend_django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-allauth==0.57.0
dj-rest-auth==5.0.2
django-anymail[sendgrid]==10.2
python-decouple==3.8
requests==2.31.0
Pillow==10.4.0
gunicorn==21.2.0
psycopg2-binary==2.9.11
whitenoise==6.6.0
google-cloud-vision==3.4.4
google-cloud-speech==2.21.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
dj-database-url==2.1.0
asgiref==3.7.2
python-dateutil==2.9.0.post0
reportlab==4.2.5
openpyxl==3.1.2
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
pillow-heif==0.21.0
qrcode[pil]==8.0
sentry-sdk==2.19.2
